requirements.txt
discord.py==1.7.3
Flask==2.0.2
python-dotenv==0.19.2
